+++ packages/components/src/components/link-pure/link-pure-utils.ts
@@ -1,17 +1,8 @@
-import type {
-  AlignLabel,
-  AlignLabelDeprecated,
-  LinkButtonIconName,
-  LinkTarget,
-  TextSize,
-  TypographyTextWeight,
-} from '../../types';
+import type { AlignLabel, LinkButtonIconName, LinkTarget, TextSize, TypographyTextWeight } from '../../types';
 import type { LinkAriaAttribute } from '../../utils';
 
 export type LinkPureIcon = LinkButtonIconName;
 export type LinkPureAriaAttribute = LinkAriaAttribute;
-/** @deprecated */
-export type LinkPureAlignLabelDeprecated = AlignLabelDeprecated;
 export type LinkPureAlignLabel = AlignLabel;
 export type LinkPureWeight = TypographyTextWeight;
 export type LinkPureSize = TextSize;